Perfil: Loader y refactor moviendo muchas cosas a PerfilUtils
- Muestra spinner de loading mientras carga.
- Muevo muchos métodos utilitarios a PerfilUtils para mayor orden.
- Fix: bug User no inicializado correctamente (rollback de anterior
commit).
