+++ app/Jobify/app/src/main/java/ar/fiuba/jobify/PerfilActivity.java
@@ -1,12 +1,9 @@
 package ar.fiuba.jobify;
 
-import android.app.DatePickerDialog;
-import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.content.res.ColorStateList;
 import android.graphics.Bitmap;
 import android.location.Location;
 import android.net.Uri;
@@ -15,7 +12,6 @@ import android.os.Environment;
 import android.provider.MediaStore;
 import android.support.annotation.IdRes;
 import android.support.annotation.LayoutRes;
-import android.support.annotation.NonNull;
 import android.support.design.widget.CollapsingToolbarLayout;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v4.app.DialogFragment;
@@ -24,20 +20,11 @@ import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.AlertDialog;
 import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.inputmethod.InputMethodManager;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.AutoCompleteTextView;
-import android.widget.BaseAdapter;
-import android.widget.DatePicker;
-import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.ListView;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.android.volley.AuthFailureError;
@@ -52,7 +39,6 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.List;
 
 import ar.fiuba.jobify.app_server_api.Contact;
@@ -61,8 +47,6 @@ import ar.fiuba.jobify.app_server_api.Employment;
 import ar.fiuba.jobify.app_server_api.Recommendation;
 import ar.fiuba.jobify.app_server_api.Solicitud;
 import ar.fiuba.jobify.app_server_api.User;
-import ar.fiuba.jobify.shared_server_api.JobPosition;
-import ar.fiuba.jobify.shared_server_api.SharedDataSingleton;
 import ar.fiuba.jobify.shared_server_api.Skill;
 import ar.fiuba.jobify.utils.PerfilUtils;
 
@@ -75,9 +59,6 @@ public class PerfilActivity extends NavDrawerActivity {
     private long fetchedUserID = 2;
     private User fetchedUser;
 
-    public static final int REQUEST_TAKE_PHOTO = 1;
-    public static final int REQUEST_PICK_IMAGE = 2;
-    public static final int REQUEST_GET_LOCATION = 3;
     String mCurrentPhotoPath = null;
     Uri mPhotoURI;
 
@@ -145,7 +126,7 @@ public class PerfilActivity extends NavDrawerActivity {
             });
         }
 
-        populateAutoCompleteLists();
+        PerfilUtils.populateAutoCompleteLists(this);
 
         if (fetchedUserID == connectedUserID) {
             if (fabAmigar != null) fabAmigar.setVisibility(View.GONE);
@@ -154,7 +135,8 @@ public class PerfilActivity extends NavDrawerActivity {
             if (fabEditar != null) fabEditar.setVisibility(View.VISIBLE);
         }
 
-        Utils.toggleViewVisibility(this, R.id.perfil_information_layout);
+        Utils.hideView(this, R.id.perfil_information_layout);
+        PerfilUtils.showProgress(this, true);
     }
 
     @Override
@@ -178,6 +160,7 @@ public class PerfilActivity extends NavDrawerActivity {
             boolean empezarEnModoEdicion = intent.getBooleanExtra(PERFIL_MODE_MESSAGE, false);
             if (empezarEnModoEdicion) {
 
+                PerfilUtils.showProgress(this, false);
                 Utils.showView(this, R.id.perfil_information_layout);
 
                 inEditingMode = false;
@@ -210,6 +193,10 @@ public class PerfilActivity extends NavDrawerActivity {
         } else {
             super.onBackPressed();
         }
+    }
+
+    public PerfilActivity getActivity() {
+        return this;
     }
 
     public static Context getContext() {
@@ -285,7 +272,7 @@ public class PerfilActivity extends NavDrawerActivity {
 //                                            dialog.dismiss();// Hace falta?
 
                                         } else if (options[which] == "Galería") {
-                                            dispatchChoosePictureIntent();
+                                            PerfilUtils.dispatchChoosePictureIntent(getActivity());
 //                                            dialog.dismiss();// Hace falta?
 
                                         } else {
@@ -309,7 +296,9 @@ public class PerfilActivity extends NavDrawerActivity {
             }
 
             // Cargo autocompletado de JobPositions y Skills según SharedData
-            populateAutoCompleteLists();
+            PerfilUtils.populateAutoCompleteLists(this);
+
+            final PerfilActivity activity = this;
 
             List<Skill> skillsList = new ArrayList<>();
             if (fetchedUser != null)
@@ -324,7 +313,7 @@ public class PerfilActivity extends NavDrawerActivity {
                 ib_skills.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
-                        agregarSkill();
+                        PerfilUtils.agregarSkill(activity, mSkillAdapter);
                     }
                 });
             }
@@ -342,7 +331,7 @@ public class PerfilActivity extends NavDrawerActivity {
                 ib_workHistory.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
-                        agregarEmployment();
+                        PerfilUtils.agregarEmployment(activity, mJobsAdapter);
                     }
                 });
             }
@@ -358,89 +347,6 @@ public class PerfilActivity extends NavDrawerActivity {
         inEditingMode = !inEditingMode;
     }
 
-    private boolean agregarEmployment() {
-        EditText et_company = (EditText) findViewById(R.id.text_perfil_employment_new_company);
-        AutoCompleteTextView et_position =
-                (AutoCompleteTextView) findViewById(R.id.text_perfil_employment_new_position);
-        EditText et_desde_mes =  (EditText) findViewById(R.id.perfil_employment_desde_mes);
-        EditText et_desde_anio = (EditText) findViewById(R.id.perfil_employment_desde_anio);
-        EditText et_hasta_mes =  (EditText) findViewById(R.id.perfil_employment_hasta_mes);
-        EditText et_hasta_anio = (EditText) findViewById(R.id.perfil_employment_hasta_anio);
-        // TODO: Cuidado si cambio por DatePickerDialog
-
-        if (et_company == null || et_company.length() == 0
-                || et_position == null || et_position.length() == 0
-                || et_desde_mes == null || et_desde_mes.length() == 0
-                || et_desde_anio == null || et_desde_anio.length() == 0
-                || et_hasta_mes == null || et_hasta_anio == null) {
-            if (et_position != null)
-                et_position.setError("Salvo 'hasta', ningún campo puede quedar vacío");
-            return false;
-        }
-        et_position.setError(null);
-
-        if ((et_hasta_mes.length() == 0 && et_hasta_anio.length() != 0)
-            || et_hasta_mes.length() != 0 && et_hasta_anio.length() == 0) {
-            et_position.setError("Para un trabajo actual, deje mes y año de 'hasta' vacíos");
-            return false;
-        }
-
-        // "Hasta" puede estar vacío
-        int hastaMes = (et_hasta_mes.length() == 0) ? 0 : Integer.valueOf(et_hasta_mes.getText().toString());
-        int hastaAnio = (et_hasta_anio.length() == 0) ? 0 : Integer.valueOf(et_hasta_anio.getText().toString());
-
-        try {
-            Employment nuevoEmployment = Employment.create(this,
-                            et_company.getText().toString(),
-                            et_position.getText().toString(),
-                            Integer.valueOf(et_desde_mes.getText().toString()),
-                            Integer.valueOf(et_desde_anio.getText().toString()),
-                            hastaMes, hastaAnio);
-            if (nuevoEmployment == null) return false;
-            if (!mJobsAdapter.add(nuevoEmployment, true)) {
-                Toast.makeText(getContext(), "Empleo ya listado", Toast.LENGTH_LONG)
-                        .show();
-                return false;
-            }
-            et_company.setText(null);
-            et_position.setText(null);
-            et_desde_mes.setText(null);
-            et_desde_anio.setText(null);
-            et_hasta_mes.setText(null);
-            et_hasta_anio.setText(null);
-            mJobsAdapter.notifyDataSetChanged();
-
-        } catch (IllegalArgumentException ex) {
-            et_position.setError("Valor inválido para " + ex.getMessage());
-            et_position.requestFocus();
-            return false;
-        }
-        return true;
-    }
-
-    private boolean agregarSkill() {
-        EditText et_skill = (EditText) findViewById(R.id.text_perfil_skill_new);
-        if (et_skill == null || et_skill.length() == 0) return false;
-
-        try {
-            Skill nuevoSkill = Skill.create(this, et_skill.getText().toString());
-            if (nuevoSkill == null) return false;
-            if (!mSkillAdapter.add(nuevoSkill, true)) {
-                Toast.makeText(getContext(), "Skill ya listado", Toast.LENGTH_LONG)
-                        .show();
-                return false;
-            }
-            et_skill.setText(null);
-            mSkillAdapter.notifyDataSetChanged();
-
-        } catch (IllegalArgumentException ex) {
-            et_skill.setError("Skill desconocido");
-            et_skill.requestFocus();
-            return false;
-        }
-
-        return true;
-    }
 
     // Si el input de algún campo es incorrecto, falla sin modificar al usuario.
     private boolean capturarInputPerfilUsuario() {
@@ -538,8 +444,7 @@ public class PerfilActivity extends NavDrawerActivity {
                     default:
                         Log.e(LOG_TAG, "This is not possible...");  // TODO: Revisar contra NONE
                 }
-
-                colorearBotonAmistad(estado);
+                PerfilUtils.colorearBotonAmistad(getActivity(), estado);
             }
         }, LOG_TAG);
     }
@@ -559,34 +464,7 @@ public class PerfilActivity extends NavDrawerActivity {
                 }, LOG_TAG);
     }
 
-    private void colorearBotonAmistad(Contact.Status estado) {
 
-        FloatingActionButton fab_amigar = (FloatingActionButton) findViewById(R.id.fab_amigar);
-        if (fab_amigar == null) {
-            Log.e(LOG_TAG, "No pude encontrar fab_amigar");
-            return;
-        }
-        ColorStateList csl;
-
-        switch (estado) {
-           case REQUESTED:
-                csl = Utils.getColorStateList(this, R.color.amigar_btn_requested);
-                break;
-            case RECEIVED:
-                csl = Utils.getColorStateList(this, R.color.amigar_btn_received);
-                break;
-            case ACTIVE:
-                csl = Utils.getColorStateList(this, R.color.amigar_btn_active);
-                break;
-            case NONE:
-                csl = Utils.getColorStateList(this, R.color.amigar_btn_none);
-                break;
-            default:
-                csl = Utils.getColorStateList(this, R.color.amigar_btn_none);
-                fab_amigar.setEnabled(false);
-        }
-        fab_amigar.setBackgroundTintList(csl);
-    }
 
     public void refreshProfileInformation(final long idFetched) {
 
@@ -612,7 +490,7 @@ public class PerfilActivity extends NavDrawerActivity {
 
     public void cargarFotoDePerfil(final long idFetched) {
 
-        // TODO / DE PRUEBA
+        // De muy poco uso
 //        if (mCurrentPhotoPath != null) {
 //            ImageView imageView = (ImageView) findViewById(R.id.perfil_image);
 //            if (imageView != null) {
@@ -639,6 +517,7 @@ public class PerfilActivity extends NavDrawerActivity {
     }
 
     private void fillProfile(User mUser) {
+        PerfilUtils.showProgress(this, false);
         Utils.showView(this, R.id.perfil_information_layout);
 
         collapsingToolbarLayout.setTitle(mUser.getFullName());
@@ -660,11 +539,11 @@ public class PerfilActivity extends NavDrawerActivity {
                     public void onResponse(JSONObject response) {
                         ContactsResponse cs = ContactsResponse.parseJSON(response.toString());
                         if (cs != null)
-                            populateContacts(cs);
+                            PerfilUtils.populateContacts(getActivity(), cs, connectedUserID);
                     }
                 }, LOG_TAG);
 
-        colorearBotonRecomendar(mUser.fueRecomendadoPor(connectedUserID));
+        PerfilUtils.colorearBotonRecomendar(this, mUser.fueRecomendadoPor(connectedUserID));
     }
 
 
@@ -692,23 +571,7 @@ public class PerfilActivity extends NavDrawerActivity {
                 }, LOG_TAG);
     }
 
-    private void dispatchChoosePictureIntent() {
-        Intent getIntent = new Intent(Intent.ACTION_GET_CONTENT);
-        getIntent.setType("image/*");
-
-        Intent pickIntent = new Intent(Intent.ACTION_PICK,
-                android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
-        pickIntent.setType("image/*");
-
-        Intent chooserIntent = Intent.createChooser(getIntent, "Select Image");
-        chooserIntent.putExtra(Intent.EXTRA_INITIAL_INTENTS, new Intent[] {pickIntent});
-
-        if (chooserIntent.resolveActivity(getPackageManager()) != null) {
-                startActivityForResult(chooserIntent, REQUEST_PICK_IMAGE);
-        }
-    }
-
-    private void dispatchTakePictureIntent() {
+    public void dispatchTakePictureIntent() {
         Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
         if (takePictureIntent.resolveActivity(getPackageManager()) != null) {
             File photoFile;
@@ -722,11 +585,11 @@ public class PerfilActivity extends NavDrawerActivity {
             }
             if (photoFile != null) {
                 mPhotoURI = FileProvider.getUriForFile(this,
-                                                    "ar.fiuba.jobify.fileprovider",
-                                                    photoFile);
+                        "ar.fiuba.jobify.fileprovider",
+                        photoFile);
 
                 takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, mPhotoURI);
-                startActivityForResult(takePictureIntent, REQUEST_TAKE_PHOTO);
+                startActivityForResult(takePictureIntent, PerfilUtils.REQUEST_TAKE_PHOTO);
             }
         }
     }
@@ -751,19 +614,20 @@ public class PerfilActivity extends NavDrawerActivity {
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
 
-        if ((requestCode == REQUEST_TAKE_PHOTO || requestCode == REQUEST_PICK_IMAGE)
+        if ( (requestCode == PerfilUtils.REQUEST_TAKE_PHOTO
+           || requestCode == PerfilUtils.REQUEST_PICK_IMAGE)
                 && resultCode == RESULT_OK) {
 
             File imageFile; // TODO: sacar repetición
 
-            if (requestCode == REQUEST_TAKE_PHOTO) {
+            if (requestCode == PerfilUtils.REQUEST_TAKE_PHOTO) {
                 imageFile = new File(mCurrentPhotoPath);
                 Log.d(LOG_TAG, "Absolute path al sacar una foto: "+mCurrentPhotoPath);
 
                 try {
                     Bitmap bitmap = MediaStore.Images.Media.getBitmap(this.getContentResolver(), mPhotoURI);
 
-                    bitmap = normalizarBitmap(bitmap);
+                    bitmap = Utils.normalizarBitmap(bitmap);
                     imageFile = guardarImagen(bitmap);
 
                 } catch (IOException ex) {
@@ -783,7 +647,7 @@ public class PerfilActivity extends NavDrawerActivity {
                     mPhotoURI = imageUri;
 
                     Bitmap bitmap = MediaStore.Images.Media.getBitmap(this.getContentResolver(), imageUri);
-                    bitmap = normalizarBitmap(bitmap);
+                    bitmap = Utils.normalizarBitmap(bitmap);
 
                     imageFile = guardarImagen(bitmap);
 
@@ -842,15 +706,9 @@ public class PerfilActivity extends NavDrawerActivity {
                 Log.w(LOG_TAG, "Archivo en " + mPhotoURI.getPath() + " no existe.");
             }
 
-        } else if (requestCode == REQUEST_GET_LOCATION && resultCode == RESULT_OK) {
+        } else if (requestCode == PerfilUtils.REQUEST_GET_LOCATION && resultCode == RESULT_OK) {
             startLocationService(null);
         }
-    }
-
-    private Bitmap normalizarBitmap(Bitmap bitmap) {
-        int newWidth = bitmap.getWidth() > 1000 ? 1000 : bitmap.getWidth(); // HARDCODEO
-        int newHeight = (int) Math.round((1.0 * bitmap.getHeight() / bitmap.getWidth()) * newWidth);
-        return Bitmap.createScaledBitmap(bitmap, newWidth, newHeight, true);
     }
 
     private File guardarImagen(Bitmap bitmap) {
@@ -878,6 +736,7 @@ public class PerfilActivity extends NavDrawerActivity {
         return imageFile;
     }
 
+
     public void startLocationService(View v) {
         Toast.makeText(this, "Actualizando ubicación", Toast.LENGTH_SHORT)
                 .show();
@@ -902,179 +761,11 @@ public class PerfilActivity extends NavDrawerActivity {
     }
 
 
-    private void populateContacts(ContactsResponse response) {
-
-        // Obtengo el estado de amistad del usuario fetched con el connected para colorear el botón
-        Contact.Status estado = response.getStatusForId(connectedUserID);
-        colorearBotonAmistad(estado);
-
-        ArrayList<Contact> contacts = response.getContactsWithStatus(Contact.Status.ACTIVE);
-
-        if (contacts.size() == 0) {
-            Utils.hideView(this, R.id.perfil_contactos_frame);
-            return;
-        } else {
-            Utils.showView(this, R.id.perfil_contactos_frame);
-        }
-
-        HorizontalListView mHLView = (HorizontalListView) findViewById(R.id.perfil_contactos_list);
-        if (mHLView != null) {
-
-            final ContactCardAdapter mAdapter = new ContactCardAdapter(contacts);
-            mHLView.setAdapter(mAdapter);
-            mHLView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-                @Override
-                public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-
-                    Contact clickedUser = (Contact) mAdapter.getItem(position);
-                    startActivity(
-                            new Intent(PerfilActivity.this, PerfilActivity.class)
-                                    .putExtra(FETCHED_USER_ID_MESSAGE, clickedUser.getId())
-                    );
-                }
-            });
-
-            TextView tv_cantContactos = (TextView) findViewById(R.id.text_perfil_cant_contactos);
-            if (tv_cantContactos != null) {
-                tv_cantContactos.setText(String.valueOf(mAdapter.getCount()));
-            }
-
-        } else {
-            Log.e(LOG_TAG, "No se encontró el gridview de contactos!");
-        }
-    }
-
-    // TODO: refactorizar?
-    private void populateAutoCompleteLists() {
-
-        try {
-            AutoCompleteTextView et_employment =
-                    (AutoCompleteTextView) findViewById(R.id.text_perfil_employment_new_position);
-            List<JobPosition> jobPositions = SharedDataSingleton.getInstance(this).getJobPositions();
-
-            if (jobPositions != null && et_employment != null) {
-                ArrayList<String> jpArray = new ArrayList<>();
-                for (JobPosition jp : jobPositions) {
-                    jpArray.add(jp.getNombre());
-                }
-                ArrayAdapter<String> employmentsAdapter = new ArrayAdapter<>(this,
-                        android.R.layout.simple_dropdown_item_1line, jpArray);
-                et_employment.setAdapter(employmentsAdapter);
-            }
-        } catch (SharedDataSingleton.NoDataException ex) {
-            Log.e(LOG_TAG, "Problemas con SS.JobPositions");
-        }
-
-        try {
-            AutoCompleteTextView et_skill =
-                    (AutoCompleteTextView) findViewById(R.id.text_perfil_skill_new);
-            List<Skill> skills = SharedDataSingleton.getInstance(this).getSkills();
-
-            if (skills != null && et_skill != null) {
-                ArrayList<String> skArray = new ArrayList<>();
-                for (Skill sk : skills) {
-                    skArray.add(sk.getNombre());
-                }
-                ArrayAdapter<String> skillsAdapter = new ArrayAdapter<>(this,
-                        android.R.layout.simple_dropdown_item_1line, skArray);
-                et_skill.setAdapter(skillsAdapter);
-            }
-        } catch (SharedDataSingleton.NoDataException ex) {
-            Log.e(LOG_TAG, "Problemas con SS.Skills");
-        }
-    }
-
-
-    private class ContactCardAdapter extends BaseAdapter {
-
-        ArrayList<Contact> mContacts;
-
-        public ContactCardAdapter(List<Contact> contactList) {
-            mContacts = new ArrayList<>(contactList);
-        }
-
-        @Override
-        public long getItemId(int position) {
-            return mContacts.get(position).getId();
-        }
-
-        @Override
-        public Object getItem(int position) {
-            return mContacts.get(position);
-        }
-
-        @Override
-        public int getCount() {
-            return mContacts.size();
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
-
-            View itemView = convertView;
-            if (itemView == null) {
-                itemView = LayoutInflater.from(getContext())
-                        .inflate(R.layout.contact_card, parent, false);
-            }
-
-            Contact contact = (Contact) getItem(position);
-            if (contact != null) {
-
-                Uri builtUri = Uri.parse(Utils.getAppServerBaseURL(getContext())).buildUpon()
-                        .appendPath(getString(R.string.get_thumbnail_path))
-                        .appendPath(Long.toString(contact.getId()))
-                        .build();
-                Utils.cargarImagenDeURLenImageView(getApplicationContext(),
-                        (ImageView) findViewById(R.id.contact_card_foto),
-                        builtUri.toString(), LOG_TAG+"CC");
-
-                TextView tv_nombre  = (TextView) itemView.findViewById(R.id.contact_card_nombre);
-                if (tv_nombre != null)
-                    tv_nombre.setText(contact.getFullName());
-
-                TextView tv_trabajo = (TextView) itemView.findViewById(R.id.contact_card_trabajo);
-                if (tv_trabajo != null)
-                    tv_trabajo.setText(contact.getCurrentJob().getOneLiner());
-            }
-
-            return itemView;
-        }
-    }
-
-
     public void showDatePickerDialog(View v) {
-        DialogFragment newFragment = new DatePickerFragment();
+        DialogFragment newFragment = new PerfilUtils.DatePickerFragment();
         newFragment.show(getSupportFragmentManager(), "datePicker");
     }
 
-    public static class DatePickerFragment extends DialogFragment
-            implements DatePickerDialog.OnDateSetListener {
-
-        PerfilActivity activity;
-
-        @Override @NonNull
-        public Dialog onCreateDialog(Bundle savedInstanceState) {
-            // No es lo más lindo del mundo...
-            // Refactor: crear interfaz que implementan, mover esto a Utils, etc.
-            activity = (PerfilActivity) getActivity();
-
-            // Use the current date pero en 1990 as the default date in the picker
-            final Calendar c = Calendar.getInstance();
-            int year = 1990;
-            int month = c.get(Calendar.MONTH);
-            int day = c.get(Calendar.DAY_OF_MONTH);
-
-            // Create a new instance of DatePickerDialog and return it
-            return new DatePickerDialog(getActivity(), this, year, month, day);
-        }
-
-        public void onDateSet(DatePicker view, int year, int month, int day) {
-            int mesBien = month + 1;
-            Utils.setTextViewText(activity, R.id.perfil_nacimiento_dia, Integer.toString(day));
-            Utils.setTextViewText(activity, R.id.perfil_nacimiento_mes, Integer.toString(mesBien));
-            Utils.setTextViewText(activity, R.id.perfil_nacimiento_anio, Integer.toString(year));
-        }
-    }
 
     private void toggleRecomendar(final long recomendadoID) {
 
@@ -1107,7 +798,7 @@ public class PerfilActivity extends NavDrawerActivity {
                                         // TODO: No chequea que el status sea 200, será inmediato?
 
                                         boolean ahoraRecomendado = !yaRecomendado;
-                                        colorearBotonRecomendar(ahoraRecomendado);
+                                        PerfilUtils.colorearBotonRecomendar(getActivity(), ahoraRecomendado);
 
                                         // TODO: Mover a R.string
                                         if (ahoraRecomendado) {
@@ -1124,22 +815,6 @@ public class PerfilActivity extends NavDrawerActivity {
                 });
     }
 
-    private void colorearBotonRecomendar(boolean recomendado) {
-
-        FloatingActionButton fab_recomendar = (FloatingActionButton) findViewById(R.id.fab_recomendar);
-        if (fab_recomendar == null) {
-            Log.e(LOG_TAG, "No pude encontrar fab_recomendar");
-            return;
-        }
-
-        ColorStateList csl;
-        if (recomendado) {
-            csl = Utils.getColorStateList(this, R.color.recomendar_btn_true);
-        } else {
-            csl = Utils.getColorStateList(this, R.color.recomendar_btn_false);
-        }
-        fab_recomendar.setBackgroundTintList(csl);
-    }
 
     public void irAConversacion(View v) {
         Utils.iniciarConversacionActivity(this, fetchedUserID);
