+++ app/Jobify/app/src/main/java/ar/fiuba/jobify/shared_server_api/SharedDataSingleton.java
@@ -20,6 +20,7 @@ public class SharedDataSingleton {
     SkillsResponse mSkillsResponse;
     static Activity mActivity;
 
+
     private SharedDataSingleton(Activity activity) {
         mActivity = activity;
