+++ app/Jobify/app/src/main/res/layout/activity_perfil.xml
@@ -83,4 +83,15 @@
         android:visibility="gone"
         app:layout_anchorGravity="bottom|end" />
 
+    <!-- Login progress -->
+    <ProgressBar
+        android:id="@+id/perfil_progress"
+        style="?android:attr/progressBarStyleLarge"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="8dp"
+        android:layout_marginTop="200dp"
+        android:layout_gravity="center_horizontal"
+        android:visibility="gone" />
+
 </android.support.design.widget.CoordinatorLayout>
