+++ app/Jobify/app/src/main/res/layout/content_user_list.xml
@@ -15,7 +15,7 @@
 
     <!-- Login progress -->
     <ProgressBar
-        android:id="@+id/search_progress"
+        android:id="@+id/busqueda_progress"
         style="?android:attr/progressBarStyleLarge"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
